python-scraper/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pymongo>=4.6.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
httpx>=0.25.0
python-multipart>=0.0.6
scikit-learn>=1.3.0
spacy>=3.7.0
pandas>=2.1.0
numpy>=1.26.0
apscheduler>=3.10.0
joblib>=1.3.0
